DEMOS-684 move styles to local usage
