+++ client/src/components/application/phases/ConceptPhase.tsx
@@ -1,9 +1,7 @@
 import React, { useState } from "react";
 
-import {
-  ApplicationUploadSection,
-  STYLES,
-} from "components/application/phases/sections";
+import { tw } from "tags/tw";
+import { ApplicationUploadSection } from "components/application/phases/sections";
 import { Button, SecondaryButton } from "components/button";
 import { ConceptPreSubmissionUploadDialog } from "components/dialog/document/ConceptPreSubmissionUploadDialog";
 import { ChevronRightIcon } from "components/icons";
@@ -13,6 +11,19 @@ import { DocumentTableDocument } from "components/table/tables/DocumentTable";
 import { isLocalDevelopment } from "config/env";
 
 import { gql, useMutation } from "@apollo/client";
+
+export const STYLES = {
+  pane: tw`bg-white p-8`,
+  grid: tw`relative grid grid-cols-2 gap-10`,
+  divider: tw`pointer-events-none absolute left-1/2 top-0 h-full border-l border-border-subtle`,
+  stepEyebrow: tw`text-xs font-semibold uppercase tracking-wide text-text-placeholder mb-2`,
+  title: tw`text-xl font-semibold mb-2`,
+  helper: tw`text-sm text-text-placeholder mb-2`,
+  list: tw`mt-4 space-y-3`,
+  fileRow: tw`bg-surface-secondary border border-border-fields px-3 py-2 flex items-center justify-between`,
+  fileMeta: tw`text-xs text-text-placeholder mt-0.5`,
+  actions: tw`mt-8 flex justify-end gap-3`,
+};
 
 const COMPLETE_CONCEPT_PHASE = gql`
   mutation CompleteConceptPhase($input: CompleteConceptPhaseInput!) {
