+++ client/src/components/application/phases/FederalCommentPhase.tsx
@@ -1,21 +1,31 @@
 import React, { useMemo, useState } from "react";
 import { tw } from "tags/tw";
-
-import {
-  ApplicationUploadSection,
-  STYLES,
-} from "components/application/phases/sections";
+import { ApplicationUploadSection } from "components/application/phases/sections";
 import { ExitIcon, WarningIcon } from "components/icons";
 import { formatDate } from "util/formatDate";
 import { DocumentTableDocument } from "components/table/tables/DocumentTable";
 import { FederalCommentUploadDialog } from "components/dialog/document/FederalCommentUploadDialog";
 
 interface FederalCommentPhaseProps {
-  demonstrationId?: string;
+  demonstrationId: string;
   phaseStartDate: Date;
   phaseEndDate: Date
   documents?: DocumentTableDocument[];
 }
+
+export const STYLES = {
+  pane: tw`bg-white p-8`,
+  grid: tw`relative grid grid-cols-2 gap-10`,
+  divider: tw`pointer-events-none absolute left-1/2 top-0 h-full border-l border-border-subtle`,
+  stepEyebrow: tw`text-xs font-semibold uppercase tracking-wide text-text-placeholder mb-2`,
+  title: tw`text-xl font-semibold mb-2`,
+  helper: tw`text-sm text-text-placeholder mb-2`,
+  list: tw`mt-4 space-y-3`,
+  fileRow: tw`bg-surface-secondary border border-border-fields px-3 py-2 flex items-center justify-between`,
+  fileMeta: tw`text-xs text-text-placeholder mt-0.5`,
+  actions: tw`mt-8 flex justify-end gap-3`,
+};
+
 
 export const FederalCommentPhase: React.FC<FederalCommentPhaseProps> = ({
   demonstrationId = "default-demo-id",
