--- client/src/components/application/phases/sections/sectionStyles.ts
@@ -1,14 +1,0 @@
-import { tw } from "tags/tw";
-
-export const STYLES = {
-  pane: tw`bg-white p-8`,
-  grid: tw`relative grid grid-cols-2 gap-10`,
-  divider: tw`pointer-events-none absolute left-1/2 top-0 h-full border-l border-border-subtle`,
-  stepEyebrow: tw`text-xs font-semibold uppercase tracking-wide text-text-placeholder mb-2`,
-  title: tw`text-xl font-semibold mb-2`,
-  helper: tw`text-sm text-text-placeholder mb-2`,
-  list: tw`mt-4 space-y-3`,
-  fileRow: tw`bg-surface-secondary border border-border-fields px-3 py-2 flex items-center justify-between`,
-  fileMeta: tw`text-xs text-text-placeholder mt-0.5`,
-  actions: tw`mt-8 flex justify-end gap-3`,
-};
